Seth's Apothecary Database Design Project
In graduate school I was tasked with creating a database for an organization from start to finish. I developed the backstory for the organization as well as the organization itself. This ended up being Seth's Apothecary, a veteran owned and operated cannabis dispensary collective out of Massachusetts. The database was made in MySQL yet used SQL by hand and not MySQL Workbench's extra tools.